+++ src/pages/vote/demo-day/index.tsx
@@ -6,17 +6,8 @@ import { useMutateRefreshing } from '@/hook/use-mutate-refreshing';
 import { mutateRefreshing } from '@/apis/auth';
 import { useEffect } from 'react';
 
-function refreshingHandler() {
-  useMutateRefreshing(mutateRefreshing);
-}
-
 export default function DemoDayVotePage() {
   const group = groupData.groups;
-
-  //page 바뀔 때, login refresh
-  useEffect(() => {
-    refreshingHandler();
-  }, []);
 
   return (
     <div className={styles.demoPage}>
